+++ src/main/scala/com/dataintuitive/viash/functionality/arguments/IntegerArgument.scala
@@ -15,35 +15,37 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.functionality.dataobjects
+package com.dataintuitive.viash.functionality.arguments
 
 import com.dataintuitive.viash.helpers.Circe.OneOrMore
 
-case class StringObject(
+case class IntegerArgument(
   name: String,
   alternatives: OneOrMore[String] = Nil,
   description: Option[String] = None,
-  example: OneOrMore[String] = Nil,
-  default: OneOrMore[String] = Nil,
+  example: OneOrMore[Int] = Nil,
+  default: OneOrMore[Int] = Nil,
   required: Boolean = false,
-  choices: List[String] = Nil,
+  choices: List[Int] = Nil,
+  min: Option[Int] = None,
+  max: Option[Int] = None,
   direction: Direction = Input,
   multiple: Boolean = false,
   multiple_sep: Char = ':',
-  `type`: String = "string"
-) extends DataObject[String] {
-  def copyDO(
+  `type`: String = "integer"
+) extends Argument[Int] {
+  def copyArg(
     `type`: String, 
     name: String, 
     alternatives: OneOrMore[String],
     description: Option[String],
-    example: OneOrMore[String],
-    default: OneOrMore[String],
+    example: OneOrMore[Int],
+    default: OneOrMore[Int],
     required: Boolean,
     direction: Direction,
     multiple: Boolean,
     multiple_sep: Char
-  ): DataObject[String] = {
-    copy(name, alternatives, description, example, default, required, this.choices, direction, multiple, multiple_sep, `type`)
+  ): Argument[Int] = {
+    copy(name, alternatives, description, example, default, required, this.choices, this.min, this.max, direction, multiple, multiple_sep, `type`)
   }
 }