+++ src/main/scala/com/dataintuitive/viash/functionality/resources/RScript.scala
@@ -18,7 +18,7 @@
 package com.dataintuitive.viash.functionality.resources
 
 import com.dataintuitive.viash.functionality._
-import com.dataintuitive.viash.functionality.dataobjects._
+import com.dataintuitive.viash.functionality.arguments._
 import com.dataintuitive.viash.wrapper.BashWrapper
 
 import java.net.URI
@@ -37,28 +37,28 @@ case class RScript(
   }
 
   def generatePlaceholder(functionality: Functionality): String = {
-    val params = functionality.allArguments.filter(d => d.direction == Input || d.isInstanceOf[FileObject])
+    val params = functionality.allArguments.filter(d => d.direction == Input || d.isInstanceOf[FileArgument])
 
     val parSet = params.map { par =>
       // val env_name = par.VIASH_PAR
       val env_name = par.viash_par_escaped("'", """\'""", """\\\'""")
 
       val parse = par match {
-        case o: BooleanObject if o.multiple =>
+        case o: BooleanArgument if o.multiple =>
           s"""as.logical(strsplit(toupper($env_name), split = '${o.multiple_sep}')[[1]])"""
-        case o: IntegerObject if o.multiple =>
+        case o: IntegerArgument if o.multiple =>
           s"""as.integer(strsplit($env_name, split = '${o.multiple_sep}')[[1]])"""
-        case o: DoubleObject if o.multiple =>
+        case o: DoubleArgument if o.multiple =>
           s"""as.numeric(strsplit($env_name, split = '${o.multiple_sep}')[[1]])"""
-        case o: FileObject if o.multiple =>
+        case o: FileArgument if o.multiple =>
           s"""strsplit($env_name, split = '${o.multiple_sep}')[[1]]"""
-        case o: StringObject if o.multiple =>
+        case o: StringArgument if o.multiple =>
           s"""strsplit($env_name, split = '${o.multiple_sep}')[[1]]"""
-        case _: BooleanObject => s"""as.logical(toupper($env_name))"""
-        case _: IntegerObject => s"""as.integer($env_name)"""
-        case _: DoubleObject => s"""as.numeric($env_name)"""
-        case _: FileObject => s"""$env_name"""
-        case _: StringObject => s"""$env_name"""
+        case _: BooleanArgument => s"""as.logical(toupper($env_name))"""
+        case _: IntegerArgument => s"""as.integer($env_name)"""
+        case _: DoubleArgument => s"""as.numeric($env_name)"""
+        case _: FileArgument => s"""$env_name"""
+        case _: StringArgument => s"""$env_name"""
       }
 
       s""""${par.plainName}" = $$VIASH_DOLLAR$$( if [ ! -z $${${par.VIASH_PAR}+x} ]; then echo "$parse"; else echo NULL; fi )"""